remove "lightning" module and move all to "trainers"
