--- torchrl/lightning/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-from .loops import RLTrainingLoop
-from .accelerators import find_device
