+++ torchrl/trainers/__init__.py
@@ -3,6 +3,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+from .accelerators import find_device
+from .loops import RLTrainingLoop
 from .trainers import (
     BatchSubSampler,
     ClearCudaCache,
@@ -18,4 +20,3 @@ from .trainers import (
     TrainerHookBase,
     UpdateWeights,
 )
-from .lightning import RLTrainingLoop
