--- torchrl/trainers/lightning.py
@@ -1,10 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-# pylint: disable=arguments-differ
-"""This module creates a `RLTrainingLoop` base class for integration with `lightning`."""
-
-
-from torchrl.lightning import RLTrainingLoop
